Base

The file base.jl implements the most basic iterator infrastructure for solvers and the definitions of the types representing general IVP (initial value problem) and solvers.

General functions for solving initial value problems

ODE.solve — Function.

solve(ivp::IVP; solver=RKIntegratorAdaptive{:rk45}, opts...)

Solve the initial value problem ivp using an algorithm solver (defaults to Runge-Kutta (4,5) integrator). One can pass additional options to the solver via keyword arguments to solve (here denoted as options). The output is a Solution type (currently simply a tuple of vectors (Vector{T},Vector{Y}), where T,Y=eltype(ivp)).

source

ODE.iterate — Function.

iterate(ivp::IVP; solver=RKIntegratorAdaptive{:rk45}, opts...)

Iterate creates an iterable Problem instance from an IVP instance (specifying the math) and from a Type{AbstractSolver} (the numerical integrator). The simplest use case is

for (t,y,dy) in iterate(...)
    # do something with t, y an dy
end

If the integration interval, defined by the keyword argument tstop, is finite you can request all the results at once by calling

collect(iterate(...)) # => Vector{Tuple{T,Y,Y}}

Notes:

usually solvers require the ivp to be in a certain form, say an

ExplicitODE.

the second argument is the Type of the solver and not an instance. The instance of the solve can only be created together with the ivp as their type parameters need to match.

Input:

ivp::IVP

S::Type{AbstractSolver}

Output:

::Problem

source

Predefined types of initial value problems

ODE.AbstractIVP — Type.

AbstractIVP{T,Y}

The abstract supertype of all IVPs (initial value problems). The type parameters T and Y correspond to the types of time and state variable respectively.

source

ODE.IVP — Type.

IVP{T,Y,F,G,J} <: AbstractIVP{T,Y}

Defines the mathematical part of an IVP (initial value problem) specified in the general form:

F(t, y) = G(t, y, dy) with y(t0)= y0

Depending on the combination of the parameters this type can represent a wide range of problems, including ODE, DAE and IMEX. Nevertheless not all solvers will support any combinations of F and G. Note that not specifying G amounts to G=dy/dt.

tspan – tuple (start_t,end_t)

y0 – initial condition

F! – in-place F function F!(t,y,res). If F=0 set to nothing.

G! – in-place G function G!(t,y,dy,res). If G=dy/dt then set to nothing (or dy if the solver supports this). Can also be a mass matrix for a RHS M dy/dt

J! – in-place Jacobian function J!(t,y,dy,res).

TODO: how to fit the sparsity pattern in J?

source

ODE.ExplicitODE — Constant.

typealias ExplicitODE{T,Y,F,J} IVP{T,Y,F,Void,J}

Can be constructed by calling

ODE.ExplicitODE(t0,y0,F!;J!=jacobian))

Explicit ODE representing the problem

dy = F(t,y) with y(t0)=y0

t0, y0: initial conditions

F!: in place version of F called by F!(t,y,dy)

J!: (optional) computes J=dF/dy in place, called with J!(t,y,J)

source

ODE.ImplicitODE — Constant.

Implicit ODE representing the problem

G(t,y,dy)=0 with y(t0)=y0 and optionally y'(t0)=dy0

t0, y0: initial conditions

G!: in place version of G called by G!(res,t,y,dy), returns residual in-place in res.

J!: (optional) computes J=dF/dy+a*dF/dy' for prescribed a, called with J!(out,t,y,dy,a). Returns Jacobian in-place in out.

source

Solver architecture

ODE.AbstractSolver — Type.

AbstractSolver

The supertype of anything which can get you to a solution of a IVP. Subtypes include: AbstractIntegrators but also DenseOutput

source

ODE.AbstractIntegrator — Type.

The abstract type of the actual algorithm to solve an IVP.

source

ODE.AbstractState — Type.

AbstractState keeps the temporary data (state) for the iterator Problem{::AbstractIntegrator}.

source

ODE.Solution — Type.

Stores a solution to the ivp.

source

The fallback constructor for AbstractSolver(ivp::IVP;opts...) ensures that an error is thrown if a solver is constructed for an unsupported type of the given IVP.

Fallback functions for solvers

Base.length(::AbstractSolver)
output(::AbstractState)
Base.eltype{T,Y}(::Type{AbstractIVP{T,Y}})